requirements.txt
streamlit
openaistreamlit run app.py